Removing dparted from the installer.
* Removing the dependency of the dparted in the installer.
* Unblocking installation on bare-metal.
