+++ common/data/packages_iso.json
@@ -12,7 +12,7 @@
                 "Linux-PAM", "attr", "libcap", "systemd", "dbus",
                 "elfutils-libelf", "sqlite-autoconf", "nspr", "nss", "popt", "lua", "rpm",
                 "gptfdisk", "tar", "gzip", "openssl", "python2", "python2-libs", "python-requests",
-                "pcre", "glib", "parted", "libsigc++", "XML-Parser", "glibmm", "dparted",
+                "pcre",
                 "libgsystem", "ostree","device-mapper","device-mapper-libs","libsepol","libselinux", 
                 "db", "libsolv", "hawkey", "python-hawkey"]
 }
