+++ installer/device.py
@@ -4,33 +4,30 @@
 #
 #    Author: Mahmoud Bassiouny <mbassiouny@vmware.com>
 
-import curses
 import subprocess
 import os
-import json
-from partition import Partition
 
 class Device(object):
-    def __init__(self, model, path, size, partitions):
+    def __init__(self, model, path, size):
         self.model = model
         self.path = path
         self.size = size
-        self.partitions = partitions
 
     @staticmethod
     def refresh_devices():
-        devices_json = subprocess.check_output(['gpartedbin', 'getdevices'], stderr=open(os.devnull, 'w'))
-        devices = Device.wrap_devices_from_dict(json.loads(devices_json))
-        return devices
+        devices_list = subprocess.check_output(['lsblk', '-S', '-I', '8', '-n', '--output', 'NAME,SIZE,MODEL'], stderr=open(os.devnull, 'w'))
+        return Device.wrap_devices_from_list(devices_list)
 
     @staticmethod
-    def wrap_devices_from_dict(dct):
-        if dct['success'] == False:
-            return [];
-
+    def wrap_devices_from_list(list):
         devices = []
-        for device in dct['devices']:
+        deviceslines = list.splitlines()
+        for deviceline in deviceslines:
+            cols = deviceline.split(None, 2)
             devices.append(
-                    Device(device['model'], device['path'], device['size'], Partition.wrap_partitions_from_dict_arr(device['partitions'])))
+                    Device(cols[2] #Model
+                        , '/dev/' + cols[0] #Path
+                        , cols[1] #size
+                        ))
 
         return devices
